Add frontend, update backend, first beta'ish version

+++ backend/src/main/java/com/example/logis/services/WorkplaceService.java
@@ -4,7 +4,6 @@ import com.example.logis.data.Company;
 import com.example.logis.data.User;
 import com.example.logis.data.Workplace;
 import com.example.logis.dtos.CreateWorkplaceRequest;
-import com.example.logis.dtos.ProjectResponse;
 import com.example.logis.dtos.WorkplaceResponse;
 import com.example.logis.exceptions.ForbiddenActionException;
 import com.example.logis.exceptions.WorkplaceNotFoundException;
@@ -18,10 +17,12 @@ import java.util.List;
 public class WorkplaceService {
     private final WorkplaceRepository workplaceRepository;
     private final CompanyService companyService;
+    private final UserService userService;
 
-    public WorkplaceService(WorkplaceRepository workplaceRepository, CompanyService companyService){
+    public WorkplaceService(WorkplaceRepository workplaceRepository, CompanyService companyService, UserService userService){
         this.workplaceRepository = workplaceRepository;
         this.companyService = companyService;
+        this.userService = userService;
     }
 
     @Transactional
@@ -56,7 +57,19 @@ public class WorkplaceService {
     }
 
     @Transactional
-    public List<WorkplaceResponse> getWorkplaces(long id) {
+    public List<WorkplaceResponse> getWorkplacesByUser(User authenticatedUser) {
+        User dbUser = userService.findUser(authenticatedUser.getId());
+        return dbUser.getCompany().getWorkplaces().stream()
+                .map(workplace -> new WorkplaceResponse(
+                        workplace.getId(),
+                        workplace.getName(),
+                        workplace.getLocation(),
+                        workplace.getCompany().getId()
+                )).toList();
+    }
+
+    @Transactional
+    public List<WorkplaceResponse> getWorkplacesByCompanyId(long id) {
         return companyService.findCompany(id).getWorkplaces().stream()
                 .map(workplace -> new WorkplaceResponse(
                         workplace.getId(),
@@ -66,7 +79,7 @@ public class WorkplaceService {
                 )).toList();
     }
 
-    public WorkplaceResponse getWorkplace(Long id) {
+    public WorkplaceResponse getWorkplaceByWorkplaceId(Long id) {
         Workplace workplace = findWorkplace(id);
         return new WorkplaceResponse(
                 workplace.getId(),
@@ -76,8 +89,31 @@ public class WorkplaceService {
         );
     }
 
-    public void deleteWorkplace(Long id) {
+    public WorkplaceResponse getWorkplaceByWorkplaceIdAndUser(Long id, User user) {
         Workplace workplace = findWorkplace(id);
+        if (!workplace.getCompany().getId().equals(user.getCompany().getId())) {
+            throw new ForbiddenActionException("User " + user.getId() + " is not part of the company that owns workplace " + id);
+        }
+
+        return new WorkplaceResponse(
+                workplace.getId(),
+                workplace.getName(),
+                workplace.getLocation(),
+                workplace.getCompany().getId()
+        );
+    }
+
+    public void deleteWorkplaceByWorkplaceId(Long id) {
+        Workplace workplace = findWorkplace(id);
+        workplaceRepository.delete(workplace);
+    }
+
+    @Transactional
+    public void deleteWorkplaceByWorkplaceIdAndUser(Long id, User user) {
+        Workplace workplace = findWorkplace(id);
+        if(!workplace.getCompany().getId().equals(user.getCompany().getId())){
+            throw new ForbiddenActionException("User " + user.getId() + " is not part of the company that owns workplace " + id);
+        }
         workplaceRepository.delete(workplace);
     }
 }
